upload_service/requirements.txt
django==2.0.2
djangorestframework
django-debug-toolbar
mongoengine==0.9
djongo
pymongo>=3.2.0
django-rest-framework-mongoengine
markdown
django-filter
setuptools
Pillow
django-cors-headers
django-rest-swagger==2.1.2
jsonfield
../avengers_django_models/avengers_django_models_package/dist/avengers_django_models_app-0.41.tar.gz
